+++ core/src/main/java/com/github/mongobee/dao/ChangeEntryIndexDao.java
