+++ core/src/main/java/com/github/mongobee/exception/MongobeeConfigurationException.java
