+++ core/src/main/java/com/github/mongobee/exception/MongobeeConnectionException.java
