+++ core/src/main/java/com/github/mongobee/exception/MongobeeException.java
