+++ core/src/main/java/com/github/mongobee/utils/StringUtils.java
@@ -1,0 +1,77 @@
+package com.github.mongobee.utils;
+
+
+/**
+ * Copied from springframework and remove all unneeded methods.
+ */
+public abstract class StringUtils {
+
+
+    //---------------------------------------------------------------------
+    // General convenience methods for working with Strings
+    //---------------------------------------------------------------------
+
+    /**
+     * Check that the given CharSequence is neither {@code null} nor of length 0.
+     * Note: Will return {@code true} for a CharSequence that purely consists of whitespace.
+     * <p><pre class="code">
+     * StringUtils.hasLength(null) = false
+     * StringUtils.hasLength("") = false
+     * StringUtils.hasLength(" ") = true
+     * StringUtils.hasLength("Hello") = true
+     * </pre>
+     *
+     * @param str the CharSequence to check (may be {@code null})
+     * @return {@code true} if the CharSequence is not null and has length
+     * @see #hasText(String)
+     */
+    public static boolean hasLength(CharSequence str) {
+        return (str != null && str.length() > 0);
+    }
+
+    /**
+     * Check whether the given CharSequence has actual text.
+     * More specifically, returns {@code true} if the string not {@code null},
+     * its length is greater than 0, and it contains at least one non-whitespace character.
+     * <p><pre class="code">
+     * StringUtils.hasText(null) = false
+     * StringUtils.hasText("") = false
+     * StringUtils.hasText(" ") = false
+     * StringUtils.hasText("12345") = true
+     * StringUtils.hasText(" 12345 ") = true
+     * </pre>
+     *
+     * @param str the CharSequence to check (may be {@code null})
+     * @return {@code true} if the CharSequence is not {@code null},
+     * its length is greater than 0, and it does not contain whitespace only
+     * @see Character#isWhitespace
+     */
+    public static boolean hasText(CharSequence str) {
+        if (!hasLength(str)) {
+            return false;
+        }
+        int strLen = str.length();
+        for (int i = 0; i < strLen; i++) {
+            if (!Character.isWhitespace(str.charAt(i))) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    /**
+     * Check whether the given String has actual text.
+     * More specifically, returns {@code true} if the string not {@code null},
+     * its length is greater than 0, and it contains at least one non-whitespace character.
+     *
+     * @param str the String to check (may be {@code null})
+     * @return {@code true} if the String is not {@code null}, its length is
+     * greater than 0, and it does not contain whitespace only
+     * @see #hasText(CharSequence)
+     */
+    public static boolean hasText(String str) {
+        return hasText((CharSequence) str);
+    }
+
+
+}
