--- src/test/resources/test.json
@@ -1,1 +1,0 @@
-{"test":"test"}
